requirements.txt
mysql-connector-python
numpy
matplotlib==3.9.0
python_dotenv
tkPDFViewer
